feat: refactor into useGitHubStats hook
